Adding search field to the landing page to find if a house exists
